% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{This is the F1 function}
\usage{
f1(text = "default")
}
\arguments{
\item{text}{(character) the text to paste with "F1". Default = "default".}
}
\value{
A character vector of lenght 1
}
\description{
This is the F1 function
}
\examples{
f1 <- f1(text = "simple")
}
